[breaking] Update to 1.5.x API (closes #6)
This commit updates the plugin to work with NodeBB 1.5.x. They changed their
API to depend on the published name of the module. This means the plugin
could not continue using "sso-discord" internally and needs to instead
use "sso-discord-alt" internally. The result is that prior plugin users
_must_ disable and uninstall the plugin prior to upgrading. Once the plugin
settings are updated with the correct Discord API keys, the plugin will work
as if nothing has changed.